src/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
grpcio==1.59.3
grpcio-tools==1.59.3
python-multipart==0.0.6
python-dotenv==1.0.0
prometheus_client
psutil